adds if statement to prevent "your location" being saved in local storage

+++ server/angular/services/localstorage-service.js
@@ -70,14 +70,15 @@
 
                         if (postcode !== address && 
                             street !== address &&
-                            uprn !== address) {
+                            uprn !== address &&
+                            address !== "your location") {
 
                             localStorageService.remove("userLocation");
                             localStorageService.remove("USER-LOCATION");
                             
                             localStorageService.set("S-USER-LOCATION", address);
                         }
-                    } else {
+                    } else if (address !== "your location") {
                         localStorageService.set("S-USER-LOCATION", address);
                     }
                 }
